+++ clx_invoice_policy/models/res_partner.py
@@ -172,9 +172,6 @@ class Partner(models.Model):
         elif self._context.get('generate_invoice_date_range'):
             start_date = self._context.get('start_date')
             end_date = self._context.get('end_date')
-            if self._context.get('regenerate_invoice', False):
-                so_lines = lines.filtered(
-                    lambda x: x.cancel_invoice_start_date >= start_date and x.cancel_invoice_end_date >= end_date)
 
             prepared_lines = [line.with_context({
                 'advance': True,