+++ clx_task_management/models/__init__.py
@@ -11,3 +11,4 @@ from . import sale_order
 from . import res_config_settings
 from . import product_category
 from . import request_form_attachments
+from . import sub_task_project